Added PCIe ECAM mapping and switched the PCIe enumeration path to happen from inside the first proper kernel thread. Also fixed the round robin scheduler to keep executing the current thread if there are no threads in the run queue but the current thread remains runnable.

+++ Justfile
@@ -41,5 +41,5 @@ clean:
     rm -rf {{image_dir}}
 
 distclean: clean
-    rm Cargo.lock
-    rm log*.txt
+    if [ -f Cargo.lock ]; then rm Cargo.lock; fi
+    if [ -d logs ]; then rm logs/*.txt; fi
